pyproject.toml
[project]
name = "panel-graphic-walker"
version = "0.3.2"
description = "A project providing a Graphic Walker Pane for use with HoloViz Panel."
readme = "README.md"
authors = [
    { name = "Philipp Rudiger", email = "philipp.jfr@gmail.com" }
]
requires-python = ">=3.9"
dependencies = ["panel>=1.5.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "jedi-language-server;sys_platform == 'linux'",
    "mypy",
    "pandas-stubs",
    "pre-commit",
    "pytest",
    "ruff",
    "watchfiles",
]
tests = [
  "duckdb",
  "gw-dsl-parser",
  "pygwalker",
  "pytest",
  "pytest-asyncio"
]
examples = [
  "duckdb",
  "fastparquet",
  "gw-dsl-parser",
  "pygwalker",
]
server = [
  "duckdb",
  "gw-dsl-parser",
  "pygwalker",
]

[tool.hatch.build.targets.wheel]
packages = ["src/panel_gwalker"]

[[tool.mypy.overrides]]
module = "param.*,pygwalker.*,gw_dsl_parser.*"
ignore_missing_imports = true
